+++ skymapper/skymapper.py
@@ -26,288 +26,6 @@ def extrap(x, xp, yp):
     y[x_ > xp_[-1]] = yp_[-1] + (x_[x_ > xp_[-1]] -xp_[-1])*(yp_[-1] - yp_[-2])/(xp_[-1] - xp_[-2])
     return y
 
-
-class Projection(object):
-
-    def transform(self, ra, dec):
-        """Convert RA/Dec into map coordinates
-
-        Args:
-            ra:  float or array of floats
-            dec: float or array of floats
-
-        Returns:
-            x,y with the same format as ra/dec
-        """
-        pass
-
-    def invert(self, x, y):
-        """Convert map coordinates into RA/Dec
-
-        Args:
-            x:  float or array of floats
-            y: float or array of floats
-
-        Returns:
-            RA,Dec with the same format as x/y
-        """
-        pass
-
-    def contains(self, x, y):
-        """Test if x/y is a valid set of map coordinates
-
-        Args:
-            x:  float or array of floats
-            y: float or array of floats
-
-        Returns:
-            Bool array with the same format as x/y
-        """
-        pass
-
-    def _wrapRA(self, ra):
-        ra_ = np.array([ra - self.ra_0]) * -1 # inverse for RA
-        # check that ra_ is between -180 and 180 deg
-        ra_[ra_ < -180 ] += 360
-        ra_[ra_ > 180 ] -= 360
-        return ra_[0]
-
-
-class AlbersEqualAreaConic(Projection):
-    def __init__(self, ra_0, dec_0, dec_1, dec_2):
-        """Albers Equal-Area projection.
-
-        AEA is a conic projection with an origin along the lines connecting
-        the poles. It preserves relative area, but is not conformal,
-        perspective or equistant.
-
-        Its preferred use of for areas with predominant east-west extent
-        at moderate latitudes.
-
-        As a conic projection, it depends on two standard parallels, i.e.
-        intersections of the cone with the sphere. To minimize scale variations,
-        these standard parallels should be chosen as small as possible while
-        spanning the range in declinations of the data.
-
-        For details, see Snyder (1987, section 14).
-
-        Args:
-            ra_0: RA that maps onto x = 0
-            dec_0: Dec that maps onto y = 0
-            dec_1: lower standard parallel
-            dec_2: upper standard parallel (must not be -dec_1)
-        """
-        self.ra_0 = ra_0
-        self.dec_0 = dec_0
-        self.dec_1 = dec_1
-        self.dec_2 = dec_2
-
-        # Snyder 1987, eq. 14-3 to 14-6.
-        self.n = (np.sin(dec_1 * DEG2RAD) + np.sin(dec_2 * DEG2RAD)) / 2
-        self.C = np.cos(dec_1 * DEG2RAD)**2 + 2 * self.n * np.sin(dec_1 * DEG2RAD)
-        self.rho_0 = self._rho(dec_0)
-
-    def _rho(self, dec):
-        return np.sqrt(self.C - 2 * self.n * np.sin(dec * DEG2RAD)) / self.n
-
-    def transform(self, ra, dec):
-        ra_ = self._wrapRA(ra)
-        # Snyder 1987, eq 14-1 to 14-4
-        theta = self.n * ra_
-        rho = self._rho(dec)
-        return rho*np.sin(theta * DEG2RAD), self.rho_0 - rho*np.cos(theta * DEG2RAD)
-
-    def contains(self, x, y):
-        rho = np.sqrt(x**2 + (self.rho_0 - y)**2)
-        inside = np.abs((self.C - (rho * self.n)**2)/(2*self.n)) <= 1
-        if self.n >= 0:
-            theta = np.arctan2(x, self.rho_0 - y) / DEG2RAD
-        else:
-            theta = np.arctan2(-x, -(self.rho_0 - y)) / DEG2RAD
-        wedge = np.abs(theta) < np.abs(self.n*180)
-        return inside & wedge
-
-    def invert(self, x, y):
-        # ra/dec actually x/y
-        # Snyder 1987, eq 14-8 to 14-11
-        rho = np.sqrt(x**2 + (self.rho_0 - y)**2)
-        if self.n >= 0:
-            theta = np.arctan2(x, self.rho_0 - y) / DEG2RAD
-        else:
-            theta = np.arctan2(-x, -(self.rho_0 - y)) / DEG2RAD
-        return self.ra_0 - theta/self.n, np.arcsin((self.C - (rho * self.n)**2)/(2*self.n)) / DEG2RAD
-
-    def __repr__(self):
-        return "AlbersEqualArea(%r, %r, %r, %r)" % (self.ra_0, self.dec_0, self.dec_1, self.dec_2)
-
-class LambertConformalConic(Projection):
-    def __init__(self, ra_0, dec_0, dec_1, dec_2):
-        """Lambert Conformal conic projection.
-
-        LCC is a conic projection with an origin along the lines connecting
-        the poles. It preserves angles, but is not equal-area,
-        perspective or equistant.
-
-        Its preferred use of for areas with predominant east-west extent
-        at higher latitudes.
-
-        As a conic projection, it depends on two standard parallels, i.e.
-        intersections of the cone with the sphere. To minimize scale variations,
-        these standard parallels should be chosen as small as possible while
-        spanning the range in declinations of the data.
-
-        For details, see Snyder (1987, section 15).
-
-        Args:
-            ra_0: RA that maps onto x = 0
-            dec_0: Dec that maps onto y = 0
-            dec_1: lower standard parallel
-            dec_2: upper standard parallel (must not be -dec_1)
-        """
-        self.ra_0 = ra_0
-        self.dec_0 = dec_0
-        self.dec_1 = dec_1
-        self.dec_2 = dec_2
-
-        # Snyder 1987, eq. 14-1, 14-2 and 15-1 to 15-3.
-        self.dec_max = 89.99
-
-        dec_1 *= DEG2RAD
-        dec_2 *= DEG2RAD
-        self.n = np.log(np.cos(dec_1)/np.cos(dec_2)) / \
-        (np.log(np.tan(np.pi/4 + dec_2/2)/np.tan(np.pi/4 + dec_1/2)))
-        self.F = np.cos(dec_1)*(np.tan(np.pi/4 + dec_1/2)**self.n)/self.n
-        self.rho_0 = self._rho(dec_0)
-
-    def _rho(self, dec):
-        # check that dec is inside of -dec_max .. dec_max
-        dec_ = np.array([dec], dtype='f8')
-        dec_[dec_ < -self.dec_max] = -self.dec_max
-        dec_[dec_ > self.dec_max] = self.dec_max
-        return self.F / np.tan(np.pi/4 + dec_[0]/2 * DEG2RAD)**self.n
-
-    def transform(self, ra, dec):
-        ra_ = self._wrapRA(ra)
-        theta = self.n * ra_
-        rho = self._rho(dec)
-        return rho*np.sin(theta * DEG2RAD), self.rho_0 - rho*np.cos(theta * DEG2RAD)
-
-    def contains(self, x, y):
-        rho = np.sqrt(x**2 + (self.rho_0 - y)**2) * np.sign(self.n)
-        inside = np.abs(rho) < max(np.abs(self._rho(self.dec_max)), np.abs(self._rho(-self.dec_max)))
-        if self.n >= 0:
-            theta = np.arctan2(x, self.rho_0 - y) / DEG2RAD
-        else:
-            theta = np.arctan2(-x, -(self.rho_0 - y)) / DEG2RAD
-        wedge = np.abs(theta) < np.abs(self.n*180)
-        return inside & wedge
-
-    def invert(self, x, y):
-        rho = np.sqrt(x**2 + (self.rho_0 - y)**2) * np.sign(self.n)
-        if self.n >= 0:
-            theta = np.arctan2(x, self.rho_0 - y) / DEG2RAD
-        else:
-            theta = np.arctan2(-x, -(self.rho_0 - y)) / DEG2RAD
-        return self.ra_0 - theta/self.n, (2 * np.arctan((self.F/rho)**(1./self.n)) - np.pi/2) / DEG2RAD
-
-    def __repr__(self):
-        return "LambertConformal(%r, %r, %r, %r)" % (self.ra_0, self.dec_0, self.dec_1, self.dec_2)
-
-
-class EquidistantConic(Projection):
-    def __init__(self, ra_0, dec_0, dec_1, dec_2):
-        """Equidistant conic projection.
-
-        Equistant conic is a projection with an origin along the lines connecting
-        the poles. It preserves distances along the map, but is not conformal,
-        perspective or equal-area.
-
-        Its preferred use of for smaller areas with predominant east-west extent
-        at moderate latitudes.
-
-        As a conic projection, it depends on two standard parallels, i.e.
-        intersections of the cone with the sphere.
-
-        For details, see Snyder (1987, section 16).
-
-        Args:
-            ra_0: RA that maps onto x = 0
-            dec_0: Dec that maps onto y = 0
-            dec_1: lower standard parallel
-            dec_2: upper standard parallel (must not be +-dec_1)
-        """
-        self.ra_0 = ra_0
-        self.dec_0 = dec_0
-        self.dec_1 = dec_1
-        self.dec_2 = dec_2
-
-        # Snyder 1987, eq. 14-3 to 14-6.
-        self.n = (np.cos(dec_1 * DEG2RAD) - np.cos(dec_2 * DEG2RAD)) / (dec_2  - dec_1) / DEG2RAD
-        self.G = np.cos(dec_1 * DEG2RAD)/self.n + (dec_1 * DEG2RAD)
-        self.rho_0 = self._rho(dec_0)
-
-    def _rho(self, dec):
-        return self.G - (dec * DEG2RAD)
-
-    def transform(self, ra, dec):
-        ra_ = self._wrapRA(ra)
-        # Snyder 1987, eq 16-1 to 16-4
-        theta = self.n * ra_
-        rho = self._rho(dec)
-        return rho*np.sin(theta * DEG2RAD), self.rho_0 - rho*np.cos(theta * DEG2RAD)
-
-    def contains(self, x, y):
-        rho = np.sqrt(x**2 + (self.rho_0 - y)**2) * np.sign(self.n)
-        rho_min = np.abs(self._rho(90))
-        rho_max = np.abs(self._rho(-90))
-        if rho_min > rho_max:
-            rho_min, rho_max = rho_max, rho_min
-        inside = (np.abs(rho) < rho_max) & (np.abs(rho) > rho_min)
-        if self.n >= 0:
-            theta = np.arctan2(x, self.rho_0 - y) / DEG2RAD
-        else:
-            theta = np.arctan2(-x, -(self.rho_0 - y)) / DEG2RAD
-        wedge = np.abs(theta) < np.abs(self.n*180)
-        return inside & wedge
-
-    def invert(self, x, y):
-        # Snyder 1987, eq 14-10 to 14-11
-        rho = np.sqrt(x**2 + (self.rho_0 - y)**2) * np.sign(self.n)
-        if self.n >= 0:
-            theta = np.arctan2(x, self.rho_0 - y) / DEG2RAD
-        else:
-            theta = np.arctan2(-x, -(self.rho_0 - y)) / DEG2RAD
-        return self.ra_0 - theta/self.n, (self.G - rho)/ DEG2RAD
-
-    def __repr__(self):
-        return "Equidistant(%r, %r, %r, %r)" % (self.ra_0, self.dec_0, self.dec_1, self.dec_2)
-
-class Hammer(Projection):
-    def __init__(self, ra_0):
-        self.ra_0 = ra_0
-
-    def transform(self, ra, dec):
-        ra_ = self._wrapRA(ra)
-        x = 2*np.sqrt(2)*np.cos(dec * DEG2RAD) * np.sin(ra_/2 * DEG2RAD)
-        y = np.sqrt(2)*np.sin(dec * DEG2RAD)
-        denom = np.sqrt(1+ np.cos(dec * DEG2RAD) * np.cos(ra_/2 * DEG2RAD))
-        return x/denom, y/denom
-
-    def invert(self, x, y):
-        dz = x*x/16 + y*y/4
-        z = np.sqrt(1- dz)
-        phi = np.arcsin(z*y) / DEG2RAD
-        lmbda = 2*np.arctan(z*x / (2*(2*z*z - 1))) / DEG2RAD
-        return self.ra_0 - lmbda, phi
-
-    def contains(self, x, y):
-        dz = x*x/16 + y*y/4
-        return dz <= 0.5
-
-    def __repr__(self):
-        return "Hammer(%r)" % self.ra_0
-
-### Map functions ###
 def degFormatter(deg):
     """Default formatter for map labels.
